Refactor drawing utilities into own class with static calls
